+++ CHANGES.md
@@ -2,11 +2,13 @@
 
 ## 2024
 
+5.0.3 text conversion works with Pillow (without background)
+
 5.0.2 copy result into clipboard under macos
 
 5.0.1 fix display image in system viewer
 
-5.0.0 use PILLOW if ImageMagick or Wand are unavailable
+5.0.0 use Pillow if ImageMagick or Wand are unavailable
 
 4.4.9 preview new when click on picture